Move methods into talk.cpp/shop.cpp
