+++ CMakeLists.txt
@@ -93,7 +93,9 @@ set(CORE_SOURCES
     random.cpp
     set_item_info.cpp
     set_option.cpp
+    shop.cpp
     std.cpp
+    talk.cpp
     talk_npc.cpp
     talk_unique.cpp
     tcg.cpp
