+++ shop.hpp
@@ -1,0 +1,13 @@
+#pragma once
+
+namespace elona
+{
+
+void shop_refresh_on_talk();
+void shop_refresh();
+void label_2266();
+void label_2267();
+void shop_load_shoptmp();
+void shop_sell_item();
+
+} // namespace elona
