+++ talk_npc.cpp
@@ -9,6 +9,7 @@
 #include "item.hpp"
 #include "item_db.hpp"
 #include "macro.hpp"
+#include "shop.hpp"
 #include "ui.hpp"
 #include "variables.hpp"
 
@@ -306,7 +307,7 @@ talk_result_t talk_npc()
         ELONA_APPEND_RESPONSE(
             46, lang(u8"免罪符を買いたい"s, u8"I want to buy an indulgence."s));
     }
-    int stat = tradecheck(tc);
+    int stat = talk_check_trade(tc);
     if (stat)
     {
         ELONA_APPEND_RESPONSE(
@@ -314,7 +315,7 @@ talk_result_t talk_npc()
     }
     if (cdata[tc].character_role == 14)
     {
-        int stat = clientguide();
+        int stat = talk_guide_quest_client();
         if (stat != 0)
         {
             for (int cnt = 0, cnt_end = (stat); cnt < cnt_end; ++cnt)
@@ -545,7 +546,7 @@ talk_result_t talk_npc()
     }
     ELONA_APPEND_RESPONSE(0, i18n::_(u8"ui", u8"bye"));
     chatesc = 1;
-    talk_window();
+    talk_window_query();
     if (chatval == 10 || chatval == 11)
     {
         if ((cdata[tc].character_role >= 1000
@@ -584,7 +585,7 @@ talk_result_t talk_npc()
     {
         invctrl = 11;
         invfile = cdata[tc].shop_store_id;
-        sell_item();
+        shop_sell_item();
         screenupdate = -1;
         update_screen();
         cs = 0;
@@ -595,7 +596,7 @@ talk_result_t talk_npc()
     {
         invctrl = 12;
         invfile = cdata[tc].shop_store_id;
-        sell_item();
+        shop_sell_item();
         cc = 0;
         screenupdate = -1;
         update_screen();
@@ -886,7 +887,7 @@ talk_result_t talk_npc()
         ELONA_APPEND_RESPONSE(1, lang(u8"挑戦する"s, u8"Alright."s));
         ELONA_APPEND_RESPONSE(0, lang(u8"やめる"s, u8"I'll pass."s));
         chatesc = 1;
-        talk_window();
+        talk_window_query();
         if (chatval != 1)
         {
             buff = lang(
@@ -953,7 +954,7 @@ talk_result_t talk_npc()
         ELONA_APPEND_RESPONSE(1, lang(u8"挑戦する"s, u8"I'll send my pet."s));
         ELONA_APPEND_RESPONSE(0, lang(u8"やめる"s, u8"I'll pass."s));
         chatesc = 1;
-        talk_window();
+        talk_window_query();
         if (chatval != 1)
         {
             buff = lang(
@@ -1062,7 +1063,7 @@ talk_result_t talk_npc()
         ELONA_APPEND_RESPONSE(
             0, lang(u8"いや、冗談です"s, u8"W-Wait! I was just kidding."s));
         chatesc = 1;
-        talk_window();
+        talk_window_query();
         if (chatval != 1)
         {
             buff = lang(u8"冷やかし"s + _ka(1), u8"You kidding? "s);
@@ -1085,7 +1086,7 @@ talk_result_t talk_npc()
             buff = lang(u8"む…中身が空っぽ"s + _dana(2), u8"Hmm! It's empty!"s);
             ELONA_APPEND_RESPONSE(0, lang(u8"しまった…"s, u8"Oops...!"s));
             chatesc = 1;
-            talk_window();
+            talk_window_query();
             modify_karma(0, -5);
         }
         else
@@ -1097,7 +1098,7 @@ talk_result_t talk_npc()
             ELONA_APPEND_RESPONSE(
                 0, lang(u8"当然のことだ"s, u8"It's nothing."s));
             chatesc = 1;
-            talk_window();
+            talk_window_query();
             modify_karma(0, 5);
             ++gdata_lost_wallet_count;
             if (gdata_lost_wallet_count >= 4)
@@ -1116,7 +1117,7 @@ talk_result_t talk_npc()
                     0,
                     lang(u8"ぎくっ"s, u8"I really found it on the street!"s));
                 chatesc = 1;
-                talk_window();
+                talk_window_query();
                 modify_karma(0, -10);
             }
         }
@@ -1150,7 +1151,7 @@ talk_result_t talk_npc()
             }
             ELONA_APPEND_RESPONSE(0, lang(u8"やめる"s, u8"Never mind."s));
             chatesc = 1;
-            talk_window();
+            talk_window_query();
             if (chatval == 1)
             {
                 snd(12);
@@ -1199,7 +1200,7 @@ talk_result_t talk_npc()
         ELONA_APPEND_RESPONSE(1, lang(u8"切る"s, u8"Yes."s));
         ELONA_APPEND_RESPONSE(0, lang(u8"やめる"s, u8"No."s));
         chatesc = 1;
-        talk_window();
+        talk_window_query();
         if (chatval == 1)
         {
             txt(lang(
@@ -1248,7 +1249,7 @@ talk_result_t talk_npc()
         }
         ELONA_APPEND_RESPONSE(0, lang(u8"やめる"s, u8"Never mind."s));
         chatesc = 1;
-        talk_window();
+        talk_window_query();
         if (chatval == 1)
         {
             txt(lang(
@@ -1282,7 +1283,7 @@ talk_result_t talk_npc()
             ELONA_APPEND_RESPONSE(1, lang(u8"売る"s, u8"Deal."s));
             ELONA_APPEND_RESPONSE(0, lang(u8"やめる"s, u8"No way."s));
             chatesc = 1;
-            talk_window();
+            talk_window_query();
             if (chatval == 1)
             {
                 txt(lang(
@@ -1390,7 +1391,7 @@ talk_result_t talk_npc()
         ELONA_APPEND_RESPONSE(1, lang(u8"切る"s, u8"Yes."s));
         ELONA_APPEND_RESPONSE(0, lang(u8"やめる"s, u8"No."s));
         chatesc = 1;
-        talk_window();
+        talk_window_query();
         if (chatval == 1)
         {
             txt(lang(
@@ -1473,7 +1474,7 @@ talk_result_t talk_npc()
         ELONA_APPEND_RESPONSE(
             0, lang(u8"やめる"s, u8"The price is too high."s));
         chatesc = 1;
-        talk_window();
+        talk_window_query();
         if (chatval == 1)
         {
             snd(12);
@@ -1506,7 +1507,7 @@ talk_result_t talk_npc()
             }
             ELONA_APPEND_RESPONSE(0, lang(u8"やめる"s, u8"No way!"s));
             chatesc = 1;
-            talk_window();
+            talk_window_query();
             if (chatval == 1)
             {
                 snd(12);
@@ -1565,7 +1566,7 @@ talk_result_t talk_npc()
         }
         ELONA_APPEND_RESPONSE(0, lang(u8"やめる"s, u8"Some other time."s));
         chatesc = 1;
-        talk_window();
+        talk_window_query();
         if (chatval == 1)
         {
             snd(12);
@@ -1640,7 +1641,7 @@ talk_result_t talk_npc()
         ELONA_APPEND_RESPONSE(1, lang(u8"売る"s, u8"Sure, take her."s));
         ELONA_APPEND_RESPONSE(0, lang(u8"やめる"s, u8"You cold bastard."s));
         chatesc = 1;
-        talk_window();
+        talk_window_query();
         if (chatval == 1)
         {
             txt(lang(u8"パエルの母親を売った…"s, u8"You sell Pael's mom..."s));
@@ -1696,7 +1697,7 @@ talk_result_t talk_npc()
         }
         ELONA_APPEND_RESPONSE(0, lang(u8"やめる"s, u8"Another time."s));
         chatesc = 1;
-        talk_window();
+        talk_window_query();
         if (chatval == 1)
         {
             snd(12);
@@ -1726,7 +1727,7 @@ talk_result_t talk_npc()
         buff = lang(
             u8"なかなかの体つき"s + _dana() + u8"よし、買"s + _u(2),
             u8"You are...quite attractive. I'll buy you."s);
-        talk_window();
+        talk_window_query();
         if (chatval != 1)
         {
             buff = lang(u8"冷やかし"s + _ka(1), u8"You kidding? "s);
@@ -1774,7 +1775,7 @@ talk_result_t talk_npc()
                 + _kure(),
             u8"Okay sweetie, I need "s + sexvalue
                 + u8" gold pieces in front."s);
-        talk_window();
+        talk_window_query();
         if (chatval != 1)
         {
             buff = lang(u8"冷やかし"s + _ka(1), u8"You kidding? "s);
@@ -1825,7 +1826,7 @@ talk_result_t talk_npc()
             ELONA_APPEND_RESPONSE(p(cnt), mapname(p(cnt)));
         }
         buff = lang(u8"つぇｔ"s, u8"tset"s);
-        talk_window();
+        talk_window_query();
         if (chatval <= 0)
         {
             buff = lang(u8"冷やかし"s + _ka(1), u8"You kidding? "s);
@@ -1844,7 +1845,7 @@ talk_result_t talk_npc()
     }
     if (chatval >= 10000)
     {
-        clientguide();
+        talk_guide_quest_client();
         rc = rtval(chatval - 10000);
         p = direction(
             cdata[0].position.x,
